fix: router에 React.lazy 적용
